util: Warn line breaking the gem5-dist script
Once the switch node in a dist-gem5 simulation gets started it listens
the first available port starting from the initially supplied one.

To bind full system nodes to the switch, the switch logfile
is parsed for the exact port number.

This is fragile and it broke when the following line:

info: tcp_iface listening on port

changed to

build/ARM/dev/net/tcp_iface.cc:97: info: tcp_iface listening on port

This patch is fixing the problem with a more robust regex matching

Signed-off-by: Giacomo Travaglini <giacomo.travaglini@arm.com>
Change-Id: I2721b3c04653ac1e09878e80d8b1ea34ec1a0f73
Reviewed-on: https://gem5-review.googlesource.com/c/public/gem5/+/62512
Maintainer: Jason Lowe-Power <power.jg@gmail.com>
Reviewed-by: Richard Cooper <richard.cooper@arm.com>
Reviewed-by: Andreas Sandberg <andreas.sandberg@arm.com>
Tested-by: kokoro <noreply+kokoro@google.com>