tests: Remove GCC-9 compiler test
When compiling GCC-9 gem5 the gem5 object files are near double the size
than when compiling with other GCC versions. This increase in size means
we need >16GB of memory available when linking. As we do not want to
mandate >16GB systems for building gem5, we are going to drop GCC-9. The
exact cause of this bug unknown.

Change-Id: I43744d421b88b79ccb21a76badd6b525e894e973