util: Adding TargetNamedBreakpoint class for gdb.
This change adds a utility script that adds the definition of the
TargetNamedBreakpoint class. This class inherits from
gdb.Breakpoint and extends the class to allow stopping the
execution only if it can extract symbol `_name` from the context
of the breakpoint and if `_name` matches a predefined set of
exact strings or regular expressions.