dev-amdgpu: Map all ROM accesses to VGA ROM address
In X86 there is a fixed physical address containing the VGA ROM, which
is 0xc0000 - 0xdffff. When multiple GPUs are instantiated, all but the
first GPU use PCI "expansion ROM" mapped at an address decided by the
OS. In order to avoid needing to load the expansion ROM region with the
ROM data, which would require an additional input file to gem5, simply
remap all expansion ROM reads/writes to the fixed VGA ROM addresses.

This works because the VGA ROM is loaded from the gem5 disk image into
the VGA ROM region. Therefore, no additional input file is needed as it
is part of the disk image already. The only drawback is that all GPU
devices must be using the same ROM. Usually this means all GPU devices
must be the same (MI210, MI300X, etc.).