arch-arm: Expose a single TableWalker to the python world
With this patch we are renaming the 4 available Arm table
walkers (dtb_walker, itb_walker, dtb_stage2, itb_stage2)
into WalkUnits, and we are defining a single TableWalker
parent, with a single port connected to the memory
subsystem.

Every table walker flavour will go through the same port
so we are effectively shifting from a multiple table walker
design into a single table walker design (with partitioned
resources)

This patch is abandoning the rigid structure of prior design
and it is allowing the pooling of walk unit resources.

Once a walk unit finishes its jobs, it signals it to the
TableWalker by adding itself into the availableWalks set
and by awakening the TableWalker.

Once awakened, the latter will go through the list of
pending walk requests and dispatch them to available walk
units

This design is cleaning up the code by removing the currState and
savedCurrState pointers. Functional walks (generated by AT instructions)
are using new WalkUnits which are specifically reserved for
functional walks:

    walk_unit_func_s1 = Param.ArmWalkUnit(
        ArmWalkUnit(walk_type="unified"),
        "Walk Unit to be used for S1 functional walks",
    )
    walk_unit_func_s2 = Param.ArmWalkUnit(
        ArmWalkUnit(walk_type="unified", is_stage2=True),
        "Walk Unit to be used for S2 functional walks",
    )

Change-Id: I0e3d8dd6e923f398e56d916384cc09060a550434
Signed-off-by: Giacomo Travaglini <giacomo.travaglini@arm.com>