arch-arm: Add stats to track PMU events
In gem5 Arm there are two ways for extracting performance
metrics from a simulation: via the generic gem5 statistics
and via the architected PMU object.

The biggest difference between the two is that, while the PMU is visible
from guest application, gem5 statistics are transparent to the simulated
system.

This effectively means:

1) While guest application is in charge of enabling/setting the
recording activity with PMU counters, event recording for gem5 stats
happens without any intervention from the guest software.
Moreover, since the number of available PMU counters is smaller than the
number of architected events, only a fraction of those events will
be recorded at the same time.

2) Extracting PMU numbers requires the guest application to read them
and write them somewhere; either to the terminal, or to a file.
Whereas gem5 stats will be automatically dumped to a human readable file
(stats.txt) at the end of simulation, without explicit programming
from the user (as point 1)

This patch is amending the ArmPMU object to dump PMU counters as
if they were common statistics. In this way we can observe PMU
events without requiring explicit programming. We can also
observe all of them without being limited by the number of
performance counters

The way we do this is by providing every PMUEvent with a pointer
to the PMU stats. Every time an event is signalled from a probe,
the PMUEvent will increment the associated stat accordingly.

Change-Id: Iae7bf29e2a67ae8572ca40302df6ea11c1875000
Signed-off-by: Giacomo Travaglini <giacomo.travaglini@arm.com>
Reviewed-by: Richard Cooper <richard.cooper@arm.com>