sim: Add option for external memory addresses
This change adds a new parameter to the System object to allow for valid
physical addresses that are not part of the Systems' physical memory
backing store. In other words, we need a way to designate a physical
address that should go through the normal memory system (caches, etc.),
but that is not managed by the System (e.g., the addresses are managed
by another system).

This change modifies isMemAddr to check both the addresses in the
backing store (physmem) and the list of potential external addresses.