sim, util: change SIGHUP to SIGCONT for shared mem transmitter
This commit changes the signal for the shared memory transmitter,
used by the gem5 orchestrator, from SIGHUP to SIGCONT. When a gem5
process is launched, there is a short period of time before signal
handlers are registered, during which the process will use the
signals' default behaviors. SIGHUP terminates processes by default,
which we want to avoid, so we instead use SIGCONT which is ignored
if a process isn't stopped.
This change fixes the issue of the hypercall_external_signal
utility killing gem5 simulations if it sends a signal too soon after
the simulations are launched.