dev: Update MI300X model to use real firmware (#2284)
The MI300X model currently is using an MI210 VBIOS and overriding the
GFX version to run MI300X ISA. This change updates the config to use a
real MI300X VBIOS and discovery table taken from real hardware. This is
mostly changing a few IDs and adding MI300X specific MMIOs.

Major differences are multiple command processors will be seen and
allows research to experiment with different ways of partitioning the
GPUs chiplets into NUMA domains as is possible on hardware.

This requires a new disk image which will be posted on gem5-resources.