arch-arm: Add utility to ArmPMU for generating statCounters
This commit is adding the archStatCounters method to the ArmPMU.  This
utility is supposed to generate a dictionary of event {id : string}
entries to be fed to the statCounters param.

It accepts a list of strings and it matches them with their related
EventTypeId integer value (id). The method returns the generated
dictionary but it does not assigns it directly to the param.
This is because we want to be able to provide the user with
the possibility of amending the dictionary (e.g. to include
implementation defined events) before the assignment is made.

Signed-off-by: Giacomo Travaglini <giacomo.travaglini.arm.com>
Change-Id: Ic673447dc963f526ba177b6ac71cd96a74bab037
Reviewed-by: Andreas Sandberg <andreas.sandberg@arm.com>