arch-x86: further modifications and code cleanup
This commit removes more of the commented out code and also
changes fs_workload.cc in .../x86 and .../x86/linux to continue
simulation if the events to exit on kernel panic and oops can't
be added, instead of stopping simulation. It also changes
fs_workload.cc in .../x86/linux  to use the on_panic and on_oops
behavior from the parameters, instead of a hard-coded behavior.