cpu: Fix looppoint analysis v25 (#2399)
This is a PR for the issue: https://github.com/gem5/gem5/issues/2395

The new probe structure uses ```dynamic_cast``` to check
the```ProbeListener```, which caused a casting mismatch with the
LoopPoint Analysis ProbeListener, because the LoopPoint Analysis
ProbeListener uses ```StaticInstPtr``` instead of ```const
StaticInstPtr``` as part of its Arg while the Probe Point that it is
listening to, ```Commit```, uses ```const StaticInstPtr```.
This commit fixes the issue.