cpu-o3: Set DynInst timestamps regardless of O3PipeView
This patch is amending the O3 CPU pipeline + DynInst
to always annotate timestamps in the DynInst object.

In this way we decouple data production from consumption.

Getting a list of timestamps can be useful for tools
willing to extract instruction metadata, regardless from
the existing O3 PipeViewer. The overhead of adding these
annotations is likely minimal

Change-Id: Ia9bd33aeec8c889f7e84432cc992d04c80b51d6a
Signed-off-by: Giacomo Travaglini <giacomo.travaglini.arm.com>