cpu: Tage GHR out-of-bounds at rollover (#2516)
The GHR rollover of TAGE can happen in a speculative state where
multiple predictions are already in flight. In case of a misprediction
we must be able to recover the history but if we just copy the maxHist
length at rollover we will run out-of-bounce.
We fix this by adding a rollback window of 1k additional bit. This
allows more than 500 predictions (TAGE-SC-L) in flight. Additionally an
assert in the restore method is added to catch out of bounds accesses in
case this is not enough.

Signed-off-by: David Schall <david.schall@tum.de>