arch-arm: Add serializing flag to MISCREGS
We extend the MiscRegInfo enum with the MISCREG_SERIALIZING
flag, to mark Misc registers as serializing/non-serializing
on a per-register basis.
We then generate different read/write instructions accordingly.
with different StaticInstFlags

By default a MiscReg is marked as serializing.

We expect to slowly transition most of the sysregs into the
non-serializing subset

Signed-off-by: Giacomo Travaglini <giacomo.travaglini.arm.com>
Change-Id: I86973780f3b67850b5ea0e6a530f698fb175d148