cpu-simple: fix numCCRegReads/Writes for the simple CPUs
Previously, numRegReads/Writes were being initialized with
numCCRegReads/Writes and numMatRegReads/Writes in the wrong order,
based on the RegClassType enum in `src/cpu/reg_class.hh`.
This caused the numCCRegReads/Writes stats in the simple CPUs to have
values that greatly differed from their values in the other core
types. This commit fixes the order to align the stats.