python: improve error message in SimObject __setattr__ (#2269)
This CL improves the error message when the attr doesn't match the cases
above in __setattr__.

Example:
One possible scenario is that User might forget to assign the
type_signature for SignalPort:

- correct: port = SignalSourcePort("bool")("port description")
- wrong: port = SignalSourcePort("port description")

And the gem5 can sill be built and error only happens on gem5 runtime
when it is actually accessing that port, however the original error
message is quite misleading. Hence creating this CL to improve the error
message so the user is able to identify the issue easily.