dev: Temporary fix for X86 board hanging forever
The example x86-ubuntu-run.py is currently hanging out due to the new
PCI host bridge architecture (#2298). It's seems to be related to some
timing added by the PCI bus. This cause the daily tests to fail due to a
timeout.

This fix removes the timing added by the PCI bus, which means that all
timing are back to (nearly) the same as before when there was no real
PCI hierarchy.

As PCI hierarchy should add some delay at some point, this is a
temporary fix and a futur PR will be made with a better fix.