base: make RefCountingPtr's operator bool() explicit (#2729)
Make RefCountingPtr's boolean conversion operator (operator bool())
explicit, to avoid masking subtle programming mistakes. For example, if
the programmer meant to type `if (ptr1->seqNum < ptr2->seqNum)` but
accidentally typed `if (ptr1 < ptr2->seqNum)`, that mistake will pass
typechecking and will manifest as a hard-to-debug runtime error. With
this fix, the typo will just result in a compile-time typechecking
error.

Change-Id: Ibe0f20ad4ca192d6af9273518581868e0f027058