mem-ruby: Make TlmGenerator a clocked object
Prior to this patch the TlmGenerator was only capable
of scheduling a batch of transactions at a specific point in time
as can bee seen from its main python API:

tran = injectAt(payload, phase, when) [1]

While dealing with absolute numbers allows for great flexibility, it is
not very practical if a user wants to sweep the frequency of the
TlmGenerator and/or only issue a selected number of transactions
per time (they will have to manually handle this behaviour
in the python config by providing a matching tick time
to the third injectAt parameter)

To automatically provide this behaviour to the TlmGenerator, we make it
a ClockedObject.

We also rework the injectAt API by making the time parameter optional.
If it is not specified, the generator does not try to schedule the
transaction at a specific point in time. Instead it simply adds it to an
unscheduled queue of transactions.

The TlmGenerator will now be ticked according to its clock domain.
It will check every tick if there are unscheduled transactions
and will try to schedule a configurable number of those.

Adding the tick method will allow us in the future to implement some
backpressure if we deem it necessary

[1]: https://github.com/gem5/gem5/blob/stable/\
    src/mem/ruby/protocol/chi/tlm/TlmGenerator.py#L57

Change-Id: I2311ac5fd7621bd23c1b73a95a777a618cb70c3c
Signed-off-by: Giacomo Travaglini <giacomo.travaglini.arm.com>
Reviewed-by: Andreas Sandberg <andreas.sandberg@arm.com>